completed load data by category.

+++ app/src/main/java/com/sunofbeaches/taobaounion/model/domain/HomePagerContent.java
@@ -13,7 +13,7 @@ public class HomePagerContent {
      */
 
     private boolean success;
-    private int code;
+    private long code;
     private String message;
     private List<DataBean> data;
 
@@ -25,7 +25,7 @@ public class HomePagerContent {
         this.success = success;
     }
 
-    public int getCode() {
+    public long getCode() {
         return code;
     }
 
@@ -79,34 +79,34 @@ public class HomePagerContent {
          * zk_final_price : 79.9
          */
 
-        private int category_id;
+        private long category_id;
         private Object category_name;
         private String click_url;
         private String commission_rate;
-        private int coupon_amount;
+        private long coupon_amount;
         private String coupon_click_url;
         private String coupon_end_time;
         private Object coupon_info;
-        private int coupon_remain_count;
+        private long coupon_remain_count;
         private String coupon_share_url;
         private String coupon_start_fee;
         private String coupon_start_time;
-        private int coupon_total_count;
+        private long coupon_total_count;
         private String item_description;
         private long item_id;
-        private int level_one_category_id;
+        private long level_one_category_id;
         private String level_one_category_name;
         private String nick;
         private String pict_url;
-        private int seller_id;
+        private long seller_id;
         private String shop_title;
         private SmallImagesBean small_images;
         private String title;
-        private int user_type;
-        private int volume;
+        private long user_type;
+        private long volume;
         private String zk_final_price;
 
-        public int getCategory_id() {
+        public long getCategory_id() {
             return category_id;
         }
 
@@ -138,7 +138,7 @@ public class HomePagerContent {
             this.commission_rate = commission_rate;
         }
 
-        public int getCoupon_amount() {
+        public long getCoupon_amount() {
             return coupon_amount;
         }
 
@@ -170,7 +170,7 @@ public class HomePagerContent {
             this.coupon_info = coupon_info;
         }
 
-        public int getCoupon_remain_count() {
+        public long getCoupon_remain_count() {
             return coupon_remain_count;
         }
 
@@ -202,7 +202,7 @@ public class HomePagerContent {
             this.coupon_start_time = coupon_start_time;
         }
 
-        public int getCoupon_total_count() {
+        public long getCoupon_total_count() {
             return coupon_total_count;
         }
 
@@ -226,7 +226,7 @@ public class HomePagerContent {
             this.item_id = item_id;
         }
 
-        public int getLevel_one_category_id() {
+        public long getLevel_one_category_id() {
             return level_one_category_id;
         }
 
@@ -258,7 +258,7 @@ public class HomePagerContent {
             this.pict_url = pict_url;
         }
 
-        public int getSeller_id() {
+        public long getSeller_id() {
             return seller_id;
         }
 
@@ -290,7 +290,7 @@ public class HomePagerContent {
             this.title = title;
         }
 
-        public int getUser_type() {
+        public long getUser_type() {
             return user_type;
         }
 
@@ -298,7 +298,7 @@ public class HomePagerContent {
             this.user_type = user_type;
         }
 
-        public int getVolume() {
+        public long getVolume() {
             return volume;
         }
